Added an example in readme where the icon name has a hyphen in it.
Ruby 1.9 hash with a dash in a key

There are some legitimate symbols that cannot be used with the new syntax. I cannot find a reference, but it appears that a symbol name matching /[a-zA-Z_][a-zA-Z_0-9]*/ is allowed with the new syntax. If the symbol name contains special characters such as '-', you have to use the Ruby 1.8 syntax, :'my-symbol-name'

Signed-off-by: Keil Miller Jr <keilmillerjr@me.com>